+++ pages/index.js
@@ -1,6 +1,6 @@
 import Head from "next/head"
-import Header from "./comp/header"
-import Home from "./comp/home"
+import Header from "./comp/_header"
+import Home from "./comp/_home"
 
 // import app's configuration
 import config from "../appconfig.json"
