+++ pages/item/[item].js
@@ -1,0 +1,7 @@
+import { useRouter } from "next/router";
+
+export default function _MCatalog(){
+    return <>
+        <p>lewebbe</p>
+    </>
+}
